3.1-drf-intro/smart_home/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.1; python_version >= '3.10'
djangorestframework==3.15.2; python_version >= '3.8'
pillow==10.4.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
sqlparse==0.5.1; python_version >= '3.8'
tzdata==2024.1; sys_platform == 'win32'
